+++ jarbas_utils/__init__.py
@@ -17,7 +17,78 @@
 
 from threading import Thread
 from time import sleep
+import os
+from os.path import  isdir, join
 from difflib import SequenceMatcher
+from jarbas_utils.configuration import read_mycroft_config
+
+
+def get_mycroft_root():
+    paths = [
+        "/opt/venvs/mycroft-core/lib/python3.7/site-packages/",  # mark1/2
+        "/opt/venvs/mycroft-core/lib/python3.4/site-packages/ ",  # old mark1 installs
+        "/home/pi/mycroft-core"  # picroft
+    ]
+    for p in paths:
+        if isdir(join(p, "mycroft")):
+            return p
+    return None
+
+
+def resolve_resource_file(res_name):
+    """Convert a resource into an absolute filename.
+
+    Resource names are in the form: 'filename.ext'
+    or 'path/filename.ext'
+
+    The system wil look for ~/.mycroft/res_name first, and
+    if not found will look at /opt/mycroft/res_name,
+    then finally it will look for res_name in the 'mycroft/res'
+    folder of the source code package.
+
+    Example:
+    With mycroft running as the user 'bob', if you called
+        resolve_resource_file('snd/beep.wav')
+    it would return either '/home/bob/.mycroft/snd/beep.wav' or
+    '/opt/mycroft/snd/beep.wav' or '.../mycroft/res/snd/beep.wav',
+    where the '...' is replaced by the path where the package has
+    been installed.
+
+    Args:
+        res_name (str): a resource path/name
+    Returns:
+        str: path to resource or None if no resource found
+    """
+    config = read_mycroft_config()
+
+    # First look for fully qualified file (e.g. a user setting)
+    if os.path.isfile(res_name):
+        return res_name
+
+    # Now look for ~/.mycroft/res_name (in user folder)
+    filename = os.path.expanduser("~/.mycroft/" + res_name)
+    if os.path.isfile(filename):
+        return filename
+
+    # Next look for /opt/mycroft/res/res_name
+    data_dir = os.path.expanduser(config['data_dir'])
+    filename = os.path.expanduser(os.path.join(data_dir, res_name))
+    if os.path.isfile(filename):
+        return filename
+
+    # Finally look for it in the source package
+    paths = [
+        "/opt/venvs/mycroft-core/lib/python3.7/site-packages/",  # mark1/2
+        "/opt/venvs/mycroft-core/lib/python3.4/site-packages/ ",  # old mark1 installs
+        "/home/pi/mycroft-core"  # picroft
+    ]
+    for p in paths:
+        filename = os.path.join(p, 'mycroft', 'res', res_name)
+        filename = os.path.abspath(os.path.normpath(filename))
+        if os.path.isfile(filename):
+            return filename
+
+    return None  # Resource cannot be resolved
 
 
 def fuzzy_match(x, against):